refactor: Refactor lf-tree adapter to modular structure
Split lf-tree adapter implementation into separate controller, handlers, and elements modules. Removed legacy and alias files, updated imports, and migrated adapter logic to use new modular functions. This improves maintainability and aligns with foundations types.
