+++ packages/core/src/components/lf-tree/handlers.tree.ts
@@ -1,0 +1,44 @@
+import {
+  LfDataNode,
+  LfTreeAdapter,
+  LfTreeAdapterHandlers,
+  LfTreeAdapterInitializerGetters,
+} from "@lf-widgets/foundations";
+
+export const createHandlers = (
+  _getters: LfTreeAdapterInitializerGetters,
+  getAdapter: () => LfTreeAdapter,
+): LfTreeAdapterHandlers => ({
+  nodeClick: (e: Event, node: LfDataNode) => {
+    const adapter = getAdapter();
+    adapter.controller.set.selection.set(node);
+    const ci: any = adapter.controller.get.compInstance;
+    ci.onLfEvent?.(e, "click", { node });
+  },
+  nodeExpand: (e: Event, node: LfDataNode) => {
+    const adapter = getAdapter();
+    adapter.controller.set.expansion.toggle(node);
+    const ci: any = adapter.controller.get.compInstance;
+    ci.onLfEvent?.(e, "expand", { node, expansion: true });
+    ci.onLfEvent?.(e, "click", { node, expansion: true });
+  },
+  nodePointerDown: (e: Event, node: LfDataNode) => {
+    const adapter = getAdapter();
+    const ci: any = adapter.controller.get.compInstance;
+    ci.onLfEvent?.(e, "pointerdown", { node });
+  },
+  filterInput: (e: CustomEvent<any>) => {
+    const adapter = getAdapter();
+    const value = e.detail.inputValue?.toLowerCase() || "";
+    let runtime = (adapter as any)._runtime as { filterTimeout?: any };
+    if (!runtime) {
+      runtime = {};
+      (adapter as any)._runtime = runtime;
+    }
+    clearTimeout(runtime.filterTimeout);
+    runtime.filterTimeout = setTimeout(() => {
+      adapter.controller.set.filter.setValue(value);
+      adapter.controller.set.filter.apply(value);
+    }, 300);
+  },
+});
