+++ packages/core/src/components/lf-tree/lf-tree-adapter.ts
@@ -1,2 +1,29 @@
-// Public barrel for adapter; implementation lives in lf-tree-adapter.impl.ts to avoid .tsx shadow
-export * from "./lf-tree-adapter.impl";
+import {
+  LfTreeAdapter,
+  LfTreeAdapterInitializerGetters,
+  LfTreeAdapterInitializerSetters,
+  LfTreeAdapterRefs,
+} from "@lf-widgets/foundations";
+import { createGetters, createSetters } from "./controller.tree";
+import { createJsx } from "./elements.tree";
+import { createHandlers } from "./handlers.tree";
+
+export type { LfTreeAdapter } from "@lf-widgets/foundations";
+
+export const createAdapter = (
+  getters: LfTreeAdapterInitializerGetters,
+  setters: LfTreeAdapterInitializerSetters,
+  getAdapter: () => LfTreeAdapter,
+): LfTreeAdapter => ({
+  controller: {
+    get: createGetters(getters),
+    set: createSetters(setters),
+  },
+  elements: { jsx: createJsx(getAdapter), refs: createRefs() },
+  handlers: createHandlers(getters, getAdapter),
+});
+
+export const createRefs = (): LfTreeAdapterRefs => ({
+  rippleSurfaces: {},
+  filterField: null,
+});
