--- packages/core/src/components/lf-tree/lf-tree-adapter.tsx
@@ -1,2 +1,0 @@
-// .tsx shim re-exporting implementation (kept temporarily for legacy import resolution)
-export * from "./lf-tree-adapter.impl";
