--- packages/core/src/components/lf-tree/lf-tree-setters.ts
@@ -1,2 +1,0 @@
-// Legacy alias: setters merged into main adapter. Remove in future major.
-export * from "./lf-tree-adapter";
